refactor: types 폴더 파일 types 제거
